+++ mayan/apps/dynamic_search/backends/django.py
@@ -26,15 +26,6 @@ class DjangoSearchBackend(SearchBackend):
             search_query.django_query
         ).distinct()
 
-    def deindex_instance(self, *args, **kwargs):
-        """This backend doesn't remove instances."""
-
-    def index_instance(self, *args, **kwargs):
-        """
-        This backend doesn't index instances. Searches query the
-        database directly.
-        """
-
     def get_search_query(
         self, query, search_model, global_and_search=False
     ):