+++ mayan/apps/dynamic_search/backends/literals.py
@@ -1,7 +1,7 @@
 from colorful.fields import RGBColorField
-
+import elasticsearch_dsl
 import whoosh
-from whoosh import qparser  # NOQA Used to initialize the whoosh.fields module
+from whoosh import qparser  # NOQA Used to initialize the whoosh.fields module.
 
 from django.db import models
 
@@ -14,8 +14,25 @@ TERM_QUOTES = ['"', '\'']
 TERM_NEGATION_CHARACTER = '-'
 TERM_SPACE_CHARACTER = ' '
 
-TEXT_LOCK_INSTANCE_DEINDEX = 'dynamic_search_whoosh_deindex_instance'
-TEXT_LOCK_INSTANCE_INDEX = 'dynamic_search_whoosh_index_instance'
+TEXT_LOCK_INSTANCE_DEINDEX = 'dynamic_search_deindex_instance'
+TEXT_LOCK_INSTANCE_INDEX = 'dynamic_search_index_instance'
+
+# Elastic search specific.
+DJANGO_TO_ELASTIC_SEARCH_FIELD_MAP = {
+    models.AutoField: {'field': elasticsearch_dsl.field.Keyword},
+    models.BooleanField: {'field': elasticsearch_dsl.field.Keyword},
+    models.CharField: {'field': elasticsearch_dsl.field.Text},
+    models.EmailField: {'field': elasticsearch_dsl.field.Keyword},
+    models.DateTimeField: {
+        'field': elasticsearch_dsl.field.Keyword,
+        'transformation': lambda value: value.isoformat()
+    },
+    models.TextField: {'field': elasticsearch_dsl.field.Text},
+    models.UUIDField: {
+        'field': elasticsearch_dsl.field.Keyword, 'transformation': str
+    },
+    RGBColorField: {'field': elasticsearch_dsl.field.Keyword},
+}
 
 # Whoosh specific.
 DJANGO_TO_WHOOSH_FIELD_MAP = {