+++ mayan/apps/dynamic_search/tasks.py
@@ -7,12 +7,16 @@ from mayan.celery import app
 
 from .classes import SearchBackend, SearchModel
 from .exceptions import DynamicSearchRetry
+from .literals import (
+    TASK_DEINDEX_INSTANCE_MAX_RETRIES, TASK_INDEX_INSTANCE_MAX_RETRIES
+)
 
 logger = logging.getLogger(name=__name__)
 
 
 @app.task(
-    bind=True, ignore_result=True, max_retries=None, retry_backoff=True
+    bind=True, ignore_result=True,
+    max_retries=TASK_DEINDEX_INSTANCE_MAX_RETRIES, retry_backoff=True
 )
 def task_deindex_instance(self, app_label, model_name, object_id):
     logger.info('Executing')
@@ -45,7 +49,8 @@ def task_index_search_model(self, search_model_full_name):
 
 
 @app.task(
-    bind=True, ignore_result=True, max_retries=None, retry_backoff=True
+    bind=True, ignore_result=True,
+    max_retries=TASK_INDEX_INSTANCE_MAX_RETRIES, retry_backoff=True
 )
 def task_index_instance(
     self, app_label, model_name, object_id, exclude_app_label=None,
@@ -72,3 +77,16 @@ def task_index_instance(
         raise self.retry(exc=exception)
 
     logger.info('Finished')
+
+
+@app.task(ignore_result=True)
+def task_reindex_backend():
+    backend = SearchBackend.get_instance()
+    backend.reset()
+
+    for search_model in SearchModel.all():
+        task_index_search_model.apply_async(
+            kwargs={
+                'search_model_full_name': search_model.get_full_name(),
+            }
+        )